+++ contracts/java/src/org/iota/wasp/contracts/tokenregistry/types/Token.java
@@ -11,7 +11,7 @@ import org.iota.wasp.wasmlib.bytes.*;
 import org.iota.wasp.wasmlib.hashtypes.*;
 
 public class Token {
-	//@formatter:off
+    //@formatter:off
 	public long      Created;     // creation timestamp
 	public String    Description; // description what minted token represents
 	public ScAgentId MintedBy;    // original minter
@@ -21,29 +21,29 @@ public class Token {
 	public String    UserDefined; // any user defined text
 	//@formatter:on
 
-	public Token() {
-	}
+    public Token() {
+    }
 
-	public Token(byte[] bytes) {
-		BytesDecoder decode = new BytesDecoder(bytes);
-		Created = decode.Int();
-		Description = decode.String();
-		MintedBy = decode.AgentId();
-		Owner = decode.AgentId();
-		Supply = decode.Int();
-		Updated = decode.Int();
-		UserDefined = decode.String();
-	}
+    public Token(byte[] bytes) {
+        BytesDecoder decode = new BytesDecoder(bytes);
+        Created = decode.Int64();
+        Description = decode.String();
+        MintedBy = decode.AgentId();
+        Owner = decode.AgentId();
+        Supply = decode.Int64();
+        Updated = decode.Int64();
+        UserDefined = decode.String();
+    }
 
-	public byte[] toBytes() {
-		return new BytesEncoder().
-				Int(Created).
-				String(Description).
-				AgentId(MintedBy).
-				AgentId(Owner).
-				Int(Supply).
-				Int(Updated).
-				String(UserDefined).
-				Data();
-	}
+    public byte[] toBytes() {
+        return new BytesEncoder().
+                Int64(Created).
+                String(Description).
+                AgentId(MintedBy).
+                AgentId(Owner).
+                Int64(Supply).
+                Int64(Updated).
+                String(UserDefined).
+                Data();
+    }
 }
